Refactor mapper tests
