+++ react/test/mapper_x0.test.ts
@@ -1,34 +1,5 @@
-import { checkGeojson, downloadPNG, getCodeFromMainField, getErrors, toggleCustomScript, urlFromCode } from './mapper-utils'
-import { safeReload, screencap, urbanstatsFixture, downloadOrCheckString, downloadCSV } from './test_utils'
-
-export function testCode(testFn: () => TestFn, geographyKind: string, universe: string, code: string, name: string, includeGeojson: boolean = false): void {
-    const url = urlFromCode(geographyKind, universe, code)
-    urbanstatsFixture(name, url)
-
-    testFn()(name, async (t) => {
-        await t.expect(code.trim()).eql((await getCodeFromMainField()).trim())
-        await t.expect(getErrors()).eql([])
-        await toggleCustomScript(t)
-        // now in autoux mode
-        await t.expect(getErrors()).eql([])
-        await toggleCustomScript(t)
-        // now in custom mode
-        await t.expect(getErrors()).eql([])
-        await t.expect(code.trim()).eql((await getCodeFromMainField()).trim())
-        await toggleCustomScript(t)
-        // now in autoux mode
-        await t.expect(getErrors()).eql([])
-        await safeReload(t)
-        await toggleCustomScript(t)
-        // back to custom mode
-        await t.expect(code.trim()).eql((await getCodeFromMainField()).trim())
-        await screencap(t, { removeEntireMap: false })
-        if (includeGeojson) {
-            await checkGeojson(t, `mapping-geojson-${name}`)
-        }
-        await downloadPNG(t)
-    })
-}
+import { testCode, urlFromCode } from './mapper-utils'
+import { urbanstatsFixture, downloadOrCheckString, downloadCSV } from './test_utils'
 
 const codeFiltered = `
 regr = regression(y=commute_transit, x1=ln(density_pw_1km), weight=population);
