+++ reference_test_screenshots/mapper_x0/Chrome/cmap-global-opacity-2.png
