+++ reference_test_screenshots/mapper_x0/Chrome/cmaprgb-alpha-channel-1.png
