+++ reference_test_screenshots/mapper_x0/Chrome/cmaprgb-alpha-channel-and-global-opacity-2.png
