+++ reference_test_screenshots/mapper_x0/Chrome/cmaprgb-global-opacity-1.png
