+++ reference_test_screenshots/mapper_x0/Chrome/code-filtered-2.png
