+++ reference_test_screenshots/mapper_x0/Chrome/code-set-center-with-expression-1.png
