+++ reference_test_screenshots/mapper_x0/Chrome/code-with-regression-1.png
