+++ reference_test_screenshots/mapper_x0/Chrome/multiline-label-2.png
