+++ reference_test_screenshots/mapper_x0/Chrome/negative-default-value-1.png
