+++ reference_test_screenshots/mapper_x0/Chrome/pmap-global-opacity-1.png
