+++ reference_test_screenshots/mapper_x0/Chrome/point-map-1.png
