+++ reference_test_screenshots/mapper_x0/Chrome/point-map-2.png
