+++ reference_test_screenshots/mapper_x0/Chrome/rgb-map-1.png
