+++ reference_test_screenshots/mapper_x0/Chrome/rgb-map-2.png
