+++ reference_test_screenshots/mapper_x0/Chrome/translucent-outline-2.png
