+++ reference_test_screenshots/mapper_x0/Chrome/translucent-outline-custom-background-2.png
