+++ reference_test_screenshots/mapper_x0/Chrome/translucent-outline-custom-background-and-text-color-1.png
