+++ reference_test_screenshots/mapper_x0/Chrome/with-outline-2.png
